fix: 파일명 오류 수정 auth => user (#91)
